fe:re-routing the bookings
