+++ frontend/app/(pages)/bookings/payment/page.tsx
@@ -1,0 +1,15 @@
+'use client'
+import React, {useState, useEffect} from 'react'
+import style from './payment.module.css'
+
+const PaymentPage = () => {
+    return (
+        <div className={style.payment}>
+            <h1>Payment</h1>
+            <p>Payment page</p>
+
+        </div>
+    )
+}
+
+export default PaymentPage
